+++ crates/miden-tx/src/lib.rs
@@ -10,8 +10,8 @@ pub use miden_objects::transaction::TransactionInputs;
 
 mod executor;
 pub use executor::{
-    DataStore, MastForestStore, NoteAccountExecution, NoteConsumptionChecker, NoteInputsCheck,
-    TransactionExecutor,
+    DataStore, ExecutionOptions, MastForestStore, NoteAccountExecution, NoteConsumptionChecker,
+    NoteInputsCheck, TransactionExecutor,
 };
 
 pub mod host;